updated: models fixed: hilt + kapt

+++ app/src/main/java/com/nikasov/cleanarchitectureapp/domain/model/Game.kt
@@ -4,10 +4,10 @@ data class Game(
     val added: Int,
     val backgroundImage: String,
     val id: Int,
-    val metacritic: Int,
+    val metacritic: Float,
     val name: String,
     val platforms: List<Platform>,
-    val playtime: Int,
-    val rating: Int,
+    val playtime: Float,
+    val rating: Float,
     val released: String,
 )
